feat: Add dockerfile and clean up
This commit is to introduce a Dockerfile to remove the burden of end users having to install the rust compiler and allow this project to be incorporated into a future deployment with ease! Additionally, the placeholder html is being removed in favor of keeping this project more rust oriented. I plan to develop a true frontend for this project at another time, but will be seperate from this repository.